Updating scan ports field type

+++ lan_nanny/modules/models/scan_port.py
@@ -22,7 +22,7 @@ class ScanPort(Base):
             },
             {
                 'name': 'elapsed_time',
-                'type': 'int'
+                'type': 'float'
             },
             {
                 'name': 'completed',
